requirements.txt
pandas
scikit-learn
flask
tweepy
python-dotenv
plotly
kaleido